pyproject.toml
[tool.poetry]
name = "portfolio-super-python-mysql-db"
version = "0.1.0"
description = ""
authors = ["MI"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.0"
flask-mail = "^0.10.0"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
